# speed up torch.sparse_mask() cpu kernel