--- a/aten/src/ATen/native/sparse/SparseTensor.cpp
+++ b/aten/src/ATen/native/sparse/SparseTensor.cpp
@@ -405,6 +405,7 @@ SparseTensor coalesce_sparse_cpu(const SparseTensor& self) {
 // sparse_mask(D, S) -> S
 //
 // Filter Tensor D by S.indices() and output a SparseTensor.
+// D and S must share the same shape.
 // --------------------------------------------------------------------
 
 template <typename scalar_t>
@@ -456,8 +457,10 @@ SparseTensor& sparse_mask_out_cpu(SparseTensor& r, const Tensor& t, const Sparse
   }
 
   if (dim > sparse_dim) {
-    LongTensor indices = at::zeros({mask._nnz()}, mask_indices.options());
 
+    // Get a flattened sparse indices, similar to NOTE [ Flatten Sparse Indices ].
+    // Keeping this implementation because it is faster than flatten_indices()
+    LongTensor indices = at::zeros({mask._nnz()}, mask_indices.options());
     for (int64_t d = 0; d < mask.sparse_dim(); d++) {
       indices.mul_(mask.size(d));
       // This used to use a buffer but I deoptimized it