--- a/aten/src/ATen/native/sparse/cuda/SparseCUDATensor.cpp
+++ b/aten/src/ATen/native/sparse/cuda/SparseCUDATensor.cpp
@@ -30,8 +30,9 @@ SparseTensor& sparse_mask_out_cuda(SparseTensor& r, const Tensor& t, const Spars
     return r;
   }
 
+  // Get a flattened sparse indices, similar to NOTE [ Flatten Sparse Indices ].
+  // Keeping this implementation because it is faster than flatten_indices()
   LongTensor indices = at::zeros({mask._nnz()}, mask_indices.options());
-
   for (int64_t d = 0; d < mask.sparse_dim(); d++) {
     indices.mul_(mask.size(d));
     // This used to use a buffer but I deoptimized it