@ezyang 2018-11-05
> D.dim > S.sparse_dim: simply use CUDA kernel implementation

I'm not sure what you mean by this.