@weiyangfb 2018-11-05
@ezyang I meant I for code in condition: D.dim > S.sparse_dim, I just copy & paste from https://github.com/pytorch/pytorch/blob/482b1366e68cd32d3ae7671e7ae12c9488464a0a/aten/src/ATen/native/sparse/cuda/SparseCUDATensor.cpp#L33-L51. I feel bad about duplicating code. Let me know if this is not ok.